requirements.txt
supabase
openai
langchain-community
langchain-openai
pypdf
python-dotenv
